Added in all the rest of the TitleLien details to the section.
